requirements.txt
streamlit>=1.36
pandas
duckdb
plotly
sqlparse
python-dotenv
google-generativeai
